+++ src/TinyDispatcher/Bootstrap/DispatcherOperationKind.cs
@@ -1,0 +1,7 @@
+namespace TinyDispatcher.Bootstrap;
+
+public enum DispatcherOperationKind
+{
+    Command,
+    Query
+}